Kinesis auth object can now fetch creds from metadata server.
In order to emulate the behavior of various AWS SDKs, we should fetch auth
credentials from multiple sources. With this patch, the client will attempt
to fetch credentials from the environment and also attempt to get them from
the metadata server if they dont exist.

Also riding on this patch are:
1. Tweaks/edit/standardize description string of some functions
2. Prettify some text that the kinesis-cli program outputs when keys are missing