#43: Ensure that all tests are run in the project
Another issue related to the bad import was that the batchproducer tests were actually never being executed in the build.

This commit updates the test command to ensure that all test files found in the project and its subdirectories are executed appropriately.

I figured out the right command from this SO post: http://stackoverflow.com/a/16353449